1_model/1_model/deploy/requirements.txt
streamlit
ultralytics
opencv-python==4.9.0.80
Pillow
numpy